2차시작까지

+++ prefix/set1.js
@@ -54,6 +54,7 @@ var clueid43 = 0
 var clueid44 = 0
 var clueid45 = 0
 var clueid46 = 0
+var clueid47 = 0
 
 
 
@@ -77,7 +78,9 @@ module.exports = {
   const user = JSON.parse(fs.readFileSync(cfilePath, "utf-8"));
     const admin = JSON.parse(fs.readFileSync(dfilePath, "utf-8"));
     const clue = JSON.parse(fs.readFileSync(efilePath, "utf-8"));
-  try {
+ 
+ if(admin.round == 0){
+    try {
     let cluedata = {};
     
     await message.channel.bulkDelete(1)
@@ -87,8 +90,10 @@ const channel2 = client.channels.cache.get(chclue3);
 const channel3 = client.channels.cache.get(chclue4);
 const channel4 = client.channels.cache.get(chclue5);
 
-    //await message.channel.send({ files: ['./roll/Preview.mp4'] });
-    
+const channel22 = client.channels.cache.get('1024599576541397033');
+const editms = await channel22.send("준비중입니다. Loding....")////특정채널 메시지
+await editms.edit({ content: "준비중입니다. 1/10" })
+
     await channel.send('시체 : 정수리에 깊게 패인 흔적이 있다. 다량의 피가 옷과 바닥에 흘렀다.\n바닥에는 쏟아진 물건들로 엉망이다.\n생방송 라이브영상은 8시에 시작되어서 8시 15분 종료되었다.')
     await channel.send('> **윤새롬**')
     await channel.send('**010> 다이어리**')
@@ -111,6 +116,7 @@ const channel4 = client.channels.cache.get(chclue5);
     .then(sent => { // 'sent' is that message you just sent
         clueid05 = sent.id;
     })
+    await editms.edit({ content: "준비중입니다. 1/10" })
     await channel.send('**060> 편지**')
     .then(sent => { // 'sent' is that message you just sent
         clueid06 = sent.id;
@@ -129,7 +135,7 @@ const channel4 = client.channels.cache.get(chclue5);
     })
     await channel.send('-----------------------------')
   
- 
+    await editms.edit({ content: "준비중입니다. 2/10" })
     await channel1.send('> **1. 한호랑**')
     .then(sent => { // 'sent' is that message you just sent
         clueid10 = sent.id;
@@ -154,6 +160,7 @@ const channel4 = client.channels.cache.get(chclue5);
     .then(sent => { // 'sent' is that message you just sent
         clueid15 = sent.id;
     })
+    await editms.edit({ content: "준비중입니다. 3/10" })
     await channel1.send('**160> CCTV**')
     .then(sent => { // 'sent' is that message you just sent
         clueid16 = sent.id;
@@ -171,6 +178,7 @@ const channel4 = client.channels.cache.get(chclue5);
         clueid19 = sent.id;
     })
     await channel1.send('-----------------------------')
+    await editms.edit({ content: "준비중입니다. 4/10" })
 
     await channel2.send('> **2. 유수호**')
     .then(sent => { // 'sent' is that message you just sent
@@ -178,7 +186,7 @@ const channel4 = client.channels.cache.get(chclue5);
     })
     await channel2.send('**210> 수호의 유튜브 계정**')
     .then(sent => { // 'sent' is that message you just sent
-        clueid121 = sent.id;
+        clueid21 = sent.id;
     })
     await channel2.send('**220> 주머니 **')
     .then(sent => { // 'sent' is that message you just sent
@@ -192,6 +200,7 @@ const channel4 = client.channels.cache.get(chclue5);
     .then(sent => { // 'sent' is that message you just sent
         clueid24 = sent.id;
     })
+    await editms.edit({ content: "준비중입니다. 5/10" })
     await channel2.send('**250> 신체검사**')
     .then(sent => { // 'sent' is that message you just sent
         clueid25 = sent.id;
@@ -209,6 +218,7 @@ const channel4 = client.channels.cache.get(chclue5);
         clueid28 = sent.id;
     })
     await channel2.send('-----------------------------')
+    await editms.edit({ content: "준비중입니다. 6/10" })
 
     await channel3.send('> **3. 오정성**')
     .then(sent => { // 'sent' is that message you just sent
@@ -234,6 +244,7 @@ const channel4 = client.channels.cache.get(chclue5);
     .then(sent => { // 'sent' is that message you just sent
         clueid34 = sent.id;
     })
+    await editms.edit({ content: "준비중입니다. 7/10" })
     await channel3.send('**360> 증언**')
     .then(sent => { // 'sent' is that message you just sent
         clueid35 = sent.id;
@@ -242,12 +253,12 @@ const channel4 = client.channels.cache.get(chclue5);
     .then(sent => { // 'sent' is that message you just sent
         clueid36 = sent.id;
     })
-    await channel3.send('**380> 집안**')
+    await channel3.send('**380> 집 내부**')
     .then(sent => { // 'sent' is that message you just sent
         clueid37 = sent.id;
     })
     await channel3.send('-----------------------------')
-
+    await editms.edit({ content: "준비중입니다. 8/10" })
     await channel4.send('> **4. 곽편집**')
     .then(sent => { // 'sent' is that message you just sent
         clueid38 = sent.id;
@@ -272,6 +283,7 @@ const channel4 = client.channels.cache.get(chclue5);
     .then(sent => { // 'sent' is that message you just sent
         clueid43 = sent.id;
     })
+    await editms.edit({ content: "준비중입니다. 9/10" })
     await channel4.send('**460> SD카드**')
     .then(sent => { // 'sent' is that message you just sent
         clueid44 = sent.id;
@@ -284,8 +296,12 @@ const channel4 = client.channels.cache.get(chclue5);
     .then(sent => { // 'sent' is that message you just sent
         clueid46 = sent.id;
     })
+    await channel4.send('**490> 메시지**')
+    .then(sent => { // 'sent' is that message you just sent
+        clueid47 = sent.id;
+    })
     await channel4.send('-----------------------------')
-    
+    await editms.edit({ content: "준비 완료 되었습니다." })
     cluedata = {
 
         dclueid01 : clueid01,
@@ -333,15 +349,46 @@ const channel4 = client.channels.cache.get(chclue5);
         dclueid43 : clueid43,
         dclueid44 : clueid44,
         dclueid45 : clueid45,
-        dclueid46 : clueid46
+        dclueid46 : clueid46,
+        dclueid47 : clueid47
         
         }
  fs.writeFileSync(efilePath, JSON.stringify(cluedata));
  
+ database = {
+    readynum : admin.readynum,
+                class1 : admin.class1,
+                player1 : admin.player1,
+                class2 : admin.class2,
+                player2 : admin.player2,
+                class3 : admin.class3,
+                player3 : admin.player3,
+                class4 : user.class,
+                player4 : user.name,
+                class5 : admin.class5,
+                player5 : admin.player5,
+                class6 : admin.class6,
+                player6 : admin.player6,
+                playerid1 : admin.playerid1,
+                playerid2 : admin.playerid2,
+                playerid3 : admin.playerid3,
+                playerid4 : user.id,
+                playerid5 : admin.playerid5,
+                playerid6 : admin.playerid6,
+                round : admin.round + 1,
+                starttime : admin.starttime,
+                endtime : admin.endtime,
+                vpoint : admin.vpoint
+   
+ }
+fs.writeFileSync(dfilePath, JSON.stringify(database));
 } catch (error) {
     console.error(error);
     
   }
-  
+}
+else{
+    message.channel.send('더이상 사용할수 없습니다.')
+}
 
   }}
